+++ src/ToolHandler.ts
@@ -5,11 +5,9 @@
  */
 
 import type {parseArguments} from './bin/chrome-devtools-mcp-cli-options.js';
-import {logger} from './logger.js';
 import type {McpContext} from './McpContext.js';
 import type {DataFormat} from './McpResponse.js';
 import {McpResponse} from './McpResponse.js';
-import type {Mutex} from './Mutex.js';
 import {SlimMcpResponse} from './SlimMcpResponse.js';
 import {ClearcutLogger} from './telemetry/ClearcutLogger.js';
 import {bucketizeLatency} from './telemetry/transformation.js';
@@ -19,6 +17,8 @@ import type {ToolCategory} from './tools/categories.js';
 import {labels, OFF_BY_DEFAULT_CATEGORIES} from './tools/categories.js';
 import type {DefinedPageTool, ToolDefinition} from './tools/ToolDefinition.js';
 import {pageIdSchema} from './tools/ToolDefinition.js';
+import {logger} from './utils/logger.js';
+import type {Mutex} from './utils/Mutex.js';
 
 export function buildFlag(category: ToolCategory) {
   return `category${category.charAt(0).toUpperCase() + category.slice(1)}`;
