+++ src/daemon/utils.ts
@@ -9,8 +9,8 @@ import os from 'node:os';
 import path from 'node:path';
 import process from 'node:process';
 
-import {logger} from '../logger.js';
 import type {YargsOptions} from '../third_party/index.js';
+import {logger} from '../utils/logger.js';
 
 export const DAEMON_SCRIPT_PATH = path.join(import.meta.dirname, 'daemon.js');
 export const INDEX_SCRIPT_PATH = path.join(
