+++ src/trace-processing/parse.ts
@@ -4,8 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {logger} from '../logger.js';
 import {DevTools} from '../third_party/index.js';
+import {logger} from '../utils/logger.js';
 
 const engine = DevTools.TraceEngine.TraceModel.Model.createWithAllHandlers();
