+++ src/utils/polyfill.ts
@@ -5,4 +5,4 @@
  */
 
 // polyfills are now bundled with all other dependencies
-import './third_party/index.js';
+import '../third_party/index.js';
