--- src/utils/types.ts
@@ -1,10 +1,0 @@
-/**
- * @license
- * Copyright 2025 Google LLC
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export interface PaginationOptions {
-  pageSize?: number;
-  pageIdx?: number;
-}
