+++ tests/setup.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import '../src/polyfill.js';
+import '../src/utils/polyfill.js';
 
 import path from 'node:path';
 import {before, it} from 'node:test';
